fix: serperated fn to fee and balance services
