+++ src/services/networkService.ts
@@ -81,6 +81,7 @@ import {
   TxPayload,
 } from '../util/network/config/network-details.types'
 import { LiquidityPoolLayer } from 'types/earn.types'
+import balanceService from './balance.service'
 
 const ERROR_ADDRESS = '0x0000000000000000000000000000000000000000'
 const L2GasOracle = '0x420000000000000000000000000000000000000F'
@@ -746,227 +747,6 @@ class NetworkService {
     })
   }
 
-  // TODO: balance.service.ts
-  async getL1FeeBalance() {
-    try {
-      const balance = await this.L1Provider!.getBalance(this.account!)
-      return utils.formatEther(balance)
-    } catch (error) {
-      console.log('NS: getL1FeeBalance error:', error)
-      return error
-    }
-  }
-
-  // TODO: balance.service.ts
-  async getL2BalanceETH() {
-    try {
-      const balance = await this.L2Provider!.getBalance(this.account!)
-      return utils.formatEther(balance)
-    } catch (error) {
-      console.log('NS: getL2BalanceETH error:', error)
-      return error
-    }
-  }
-
-  // TODO: balance.service.ts
-  async getL2BalanceBOBA() {
-    try {
-      const ERC20Contract = new ethers.Contract(
-        this.tokenAddresses!['BOBA'].L2,
-        L2StandardERC20ABI, //any old abi will do...
-        this.provider!.getSigner()
-      )
-      const balance = await ERC20Contract.balanceOf(this.account)
-      return utils.formatEther(balance)
-    } catch (error) {
-      console.log('NS: getL2BalanceBOBA error:', error)
-      return error
-    }
-  }
-
-  // TODO: balance.service.ts
-  async getBalances() {
-    try {
-      let layer1Balances
-      let layer2Balances
-      if (this.network === Network.ETHEREUM) {
-        layer1Balances = [
-          {
-            address: this.addresses.L1_ETH_Address,
-            addressL2: this.addresses.L2_ETH_Address,
-            currency: this.addresses.L1_ETH_Address,
-            symbol: 'ETH',
-            decimals: 18,
-            balance: new BN(0),
-          },
-        ]
-
-        layer2Balances = [
-          {
-            address: this.addresses.L2_ETH_Address,
-            addressL1: this.addresses.L1_ETH_Address,
-            addressL2: this.addresses.L2_ETH_Address,
-            currency: this.addresses.L1_ETH_Address,
-            symbol: 'ETH',
-            decimals: 18,
-            balance: new BN(0),
-          },
-        ]
-      } else {
-        layer1Balances = [
-          {
-            address: this.addresses.L1_ETH_Address,
-            addressL2:
-              this.addresses['TK_L2' + networkService.L1NativeTokenSymbol],
-            currency: this.addresses.L1_ETH_Address,
-            symbol: networkService.L1NativeTokenSymbol,
-            decimals: 18,
-            balance: new BN(0),
-          },
-        ]
-
-        layer2Balances = [
-          {
-            address: this.addresses.L2_ETH_Address,
-            addressL1: this.addresses.TK_L1BOBA,
-            addressL2: this.addresses.L2_ETH_Address,
-            currency: this.addresses.TK_L1BOBA,
-            symbol: 'BOBA',
-            decimals: 18,
-            balance: new BN(0),
-          },
-        ]
-      }
-
-      // Always check ETH
-      const layer1Balance = await this.L1Provider!.getBalance(this.account!)
-      const layer2Balance = await this.L2Provider!.getBalance(this.account!)
-
-      layer1Balances[0].balance = new BN(layer1Balance.toString())
-      layer2Balances[0].balance = new BN(layer2Balance.toString())
-
-      const state = store.getState()
-      const tA = Object.values(state.tokenList)
-
-      const tokenC = new ethers.Contract(
-        this.addresses.L1_ETH_Address,
-        L1ERC20ABI,
-        this.L1Provider
-      )
-
-      const getERC20Balance = async (token, tokenAddress, layer, provider) => {
-        const balance = await tokenC
-          .attach(tokenAddress)
-          .connect(provider)
-          .balanceOf(this.account)
-        return {
-          ...token,
-          balance: new BN(balance.toString()),
-          layer,
-          address: layer === 'L1' ? token.addressL1 : token.addressL2,
-          symbol: token.symbolL1,
-        }
-      }
-
-      const getBalancePromise: any = []
-
-      tA.forEach((token: any) => {
-        if (token.addressL1 === null) {
-          return
-        }
-        if (token.addressL2 === null) {
-          return
-        }
-        if (this.network === Network.ETHEREUM) {
-          if (token.addressL1 === this.addresses.L1_ETH_Address) {
-            return
-          }
-          if (token.addressL2 === this.addresses.L2_ETH_Address) {
-            return
-          }
-        } else {
-          if (token.addressL1 === this.addresses.L1_ETH_Address) {
-            return getBalancePromise.push(
-              getERC20Balance(token, token.addressL2, 'L2', this.L2Provider)
-            )
-          }
-          if (token.addressL2 === this.addresses.L2_BOBA_Address) {
-            return getBalancePromise.push(
-              getERC20Balance(token, token.addressL1, 'L1', this.L1Provider)
-            )
-          }
-        }
-
-        if (
-          [
-            'xBOBA',
-            'WAGMIv0',
-            'WAGMIv1',
-            'WAGMIv2',
-            'WAGMIv2-Oolong',
-            'WAGMIv3',
-            'WAGMIv3-Oolong',
-            'OLO',
-          ].includes(token.symbolL1)
-        ) {
-          //there is no L1 xBOBA, WAGMIv0, WAGMIv1, WAGMIv2, WAGMIv2OLO, WAGMIv3, WAGMIv3OLO, OLO
-          getBalancePromise.push(
-            getERC20Balance(token, token.addressL2, 'L2', this.L2Provider)
-          )
-        } else {
-          getBalancePromise.push(
-            getERC20Balance(token, token.addressL1, 'L1', this.L1Provider)
-          )
-          getBalancePromise.push(
-            getERC20Balance(token, token.addressL2, 'L2', this.L2Provider)
-          )
-        }
-      })
-
-      const tokenBalances = await Promise.allSettled(getBalancePromise).then(
-        (results) =>
-          results
-            .filter((result) => {
-              switch (result.status) {
-                case 'fulfilled': {
-                  return true
-                }
-                case 'rejected': {
-                  console.log('NS: getBalances:', result.reason)
-                  return false
-                }
-              }
-            })
-            .map((result: any) => result.value)
-      )
-
-      tokenBalances.forEach((token) => {
-        if (
-          token.layer === 'L1' &&
-          token.symbol !== 'xBOBA' &&
-          token.symbol !== 'WAGMIv0' &&
-          token.symbol !== 'WAGMIv1' &&
-          token.symbol !== 'WAGMIv2' &&
-          token.symbol !== 'WAGMIv2-Oolong' &&
-          token.symbol !== 'WAGMIv3' &&
-          token.symbol !== 'WAGMIv3-Oolong'
-        ) {
-          layer1Balances.push(token)
-        } else if (token.layer === 'L2') {
-          layer2Balances.push(token)
-        }
-      })
-
-      return {
-        layer1: orderBy(layer1Balances, (i) => i.currency),
-        layer2: orderBy(layer2Balances, (i) => i.currency),
-      }
-    } catch (error) {
-      console.log('NS: getBalances error:', error)
-      return error
-    }
-  }
-
   // TODO: bridging.service.ts
   //Move ETH from L1 to L2 using the standard deposit system
   /******
@@ -1310,7 +1090,7 @@ class NetworkService {
       }
       const receipt = await this.watcher!.waitForMessageReceipt(depositTX, opts)
       const txReceipt = receipt.transactionReceipt
-      await this.getBalances()
+      await balanceService.getBalances()
       return txReceipt
     } catch (error) {
       console.log('NS: depositErc20 error:', error)
@@ -1416,174 +1196,6 @@ class NetworkService {
       return tx
     } catch (error) {
       console.log('NS: exitBOBA error:', error)
-      return error
-    }
-  }
-
-  // TODO: balance.service.ts
-  /* Estimate cost of Classical Exit to L1 */
-  async getExitCost(currencyAddress: string) {
-    try {
-      let approvalCost_BN = BigNumber.from('0')
-
-      const gasPrice = await this.L2Provider!.getGasPrice()
-      console.log('Classical exit gas price', gasPrice.toString())
-
-      if (currencyAddress !== this.addresses.L2_ETH_Address) {
-        const ERC20Contract = new ethers.Contract(
-          currencyAddress,
-          L2StandardERC20ABI, //any old abi will do...
-          this.provider!.getSigner()
-        )
-
-        const tx = await ERC20Contract.populateTransaction.approve(
-          this.addresses.DiscretionaryExitFee,
-          utils.parseEther('1.0')
-        )
-
-        const approvalGas_BN = await this.L2Provider!.estimateGas({
-          ...tx,
-          from: this.gasEstimateAccount,
-        })
-        approvalCost_BN = approvalGas_BN.mul(gasPrice)
-        console.log('Approve cost in ETH:', utils.formatEther(approvalCost_BN))
-      }
-
-      const DiscretionaryExitFeeContract = new ethers.Contract(
-        this.addresses.DiscretionaryExitFee,
-        DiscretionaryExitFeeABI,
-        this.provider!.getSigner()
-      )
-
-      const L2BillingContract = new ethers.Contract(
-        this.addresses.Proxy__BobaBillingContract,
-        L2BillingContractABI,
-        this.L2Provider
-      )
-      const exitFee = await L2BillingContract.exitFee()
-      let value = utils.parseEther('0.00001').add(exitFee)
-      if (this.networkGateway === Network.ETHEREUM) {
-        value = utils.parseEther('0.00001')
-      }
-
-      const tx2 =
-        await DiscretionaryExitFeeContract.populateTransaction.payAndWithdraw(
-          this.addresses.L2_ETH_Address,
-          utils.parseEther('0.00001'),
-          this.L1GasLimit,
-          ethers.utils.formatBytes32String(new Date().getTime().toString()),
-          { value }
-        )
-
-      const gas_BN = await this.L2Provider!.estimateGas({
-        ...tx2,
-        from: this.gasEstimateAccount,
-      })
-      console.log('Classical exit gas', gas_BN.toString())
-
-      const cost_BN = gas_BN.mul(gasPrice)
-      console.log('Classical exit cost (ETH):', utils.formatEther(cost_BN))
-
-      const totalCost = utils.formatEther(cost_BN.add(approvalCost_BN))
-      console.log('Classical exit total cost (ETH):', totalCost)
-
-      //returns total cost in ETH
-      return totalCost
-    } catch (error) {
-      return 0
-    }
-  }
-
-  /***********************************************/
-  /*****                  Fee                *****/
-  /***** Fees are reported as integers,      *****/
-  /***** where every int represents 0.1%     *****/
-  /***********************************************/
-  // TODO: balance.service.ts
-  async getL1TotalFeeRate() {
-    try {
-      const L1LPContract = new ethers.Contract(
-        this.addresses.L1LPAddress,
-        L1LiquidityPoolABI,
-        this.L1Provider
-      )
-      const [operatorFeeRate, userMinFeeRate, userMaxFeeRate] =
-        await Promise.all([
-          L1LPContract.ownerRewardFeeRate(),
-          L1LPContract.userRewardMinFeeRate(),
-          L1LPContract.userRewardMaxFeeRate(),
-        ])
-
-      const feeRateL = Number(userMinFeeRate) + Number(operatorFeeRate)
-      const feeRateH = Number(userMaxFeeRate) + Number(operatorFeeRate)
-
-      return {
-        feeMin: (feeRateL / 10).toFixed(1),
-        feeMax: (feeRateH / 10).toFixed(1),
-      }
-    } catch (error) {
-      console.log('NS: getL1TotalFeeRate error:', error)
-      return error
-    }
-  }
-  // TODO: balance.service.ts
-  async getL2TotalFeeRate() {
-    try {
-      const L2LPContract = new ethers.Contract(
-        this.addresses.L2LPAddress,
-        L2LiquidityPoolABI,
-        this.L2Provider
-      )
-      const [operatorFeeRate, userMinFeeRate, userMaxFeeRate] =
-        await Promise.all([
-          L2LPContract.ownerRewardFeeRate(),
-          L2LPContract.userRewardMinFeeRate(),
-          L2LPContract.userRewardMaxFeeRate(),
-        ])
-
-      const feeRateL = Number(userMinFeeRate) + Number(operatorFeeRate)
-      const feeRateH = Number(userMaxFeeRate) + Number(operatorFeeRate)
-
-      return {
-        feeMin: (feeRateL / 10).toFixed(1),
-        feeMax: (feeRateH / 10).toFixed(1),
-      }
-    } catch (error) {
-      console.log('NS: getL2TotalFeeRate error:', error)
-      return error
-    }
-  }
-
-  // TODO: balance.service.ts
-  async getL1UserRewardFeeRate(tokenAddress) {
-    try {
-      const L1LPContract = new ethers.Contract(
-        this.addresses.L1LPAddress,
-        L1LiquidityPoolABI,
-        this.L1Provider
-      )
-      const feeRate = await L1LPContract.getUserRewardFeeRate(tokenAddress)
-      //console.log("NS: getL1UserRewardFeeRate:", feeRate)
-      return (feeRate / 10).toFixed(1)
-    } catch (error) {
-      console.log('NS: getL1UserRewardFeeRate error:', error)
-      return error
-    }
-  }
-
-  // TODO: balance.service.ts
-  async getL2UserRewardFeeRate(tokenAddress) {
-    try {
-      const L2LPContract = new ethers.Contract(
-        this.addresses.L2LPAddress,
-        L2LiquidityPoolABI,
-        this.L2Provider
-      )
-      const feeRate = await L2LPContract.getUserRewardFeeRate(tokenAddress)
-      //console.log("NS: getL2UserRewardFeeRate:", feeRate)
-      return (feeRate / 10).toFixed(1)
-    } catch (error) {
-      console.log('NS: getL2UserRewardFeeRate error:', error)
       return error
     }
   }
